Split over-budget files for the length guard; no behavior changes
The dismiss-sync feature pushed five tracked files over their length
budgets. Extractions instead of budget refreshes:

- cmuxTests/TerminalAndGhosttyTests.swift: the dismiss-sync tests move
  to cmuxTests/NotificationDismissSyncTests.swift (wired in pbxproj).
- Sources/TerminalController.swift: the mobile notification.dismiss /
  notification.reconcile handlers and the mobile workspace-action gate
  move to Sources/TerminalController+MobileNotificationSync.swift.
- Sources/Mobile/MobileHostService.swift: mobileHostCapabilities moves
  to Sources/Mobile/MobileHostService+Capabilities.swift.
- MobileShellComposite.swift: the dismiss-sync section moves to
  MobileShellComposite+NotificationDismissSync.swift and the dogfood
  feedback round-trip to MobileShellComposite+DogfoodFeedback.swift;
  four members lose `private` (same module) for the split.
- Sources/TerminalNotificationStore.swift: the pre-existing 600-line
  NotificationSoundSettings enum moves wholesale to
  Sources/NotificationSoundSettings.swift. Its call graph is too
  entangled to split under 500 lines without ~12 access relaxations,
  so it becomes a tracked budget entry at its actual 603 lines;
  TerminalNotificationStore's budget tightens 2527 -> 2274. The TSV is
  re-sorted (it had drifted out of size order).
