+++ Packages/CmuxMobileShell/Sources/CmuxMobileShell/MobileShellComposite.swift
@@ -403,13 +403,16 @@ public final class MobileShellComposite: MobileTerminalOutputSinking {
     private let deviceRegistry: (any DeviceRegistryRefreshing)?
     private let identityProvider: (any MobileIdentityProviding)?
     private let reachability: any ReachabilityProviding
-    private let deliveredNotificationClearer: any DeliveredNotificationClearing
+    // Internal (not private): used by the dismiss-sync extension file.
+    let deliveredNotificationClearer: any DeliveredNotificationClearing
     /// Durable outbox for phone→Mac dismissals (see
     /// ``PendingNotificationDismissQueue``): ids are enqueued before the RPC is
     /// attempted, removed only on confirmed delivery, and re-flushed on every
     /// successful (re)subscribe.
-    private let pendingDismissQueue: PendingNotificationDismissQueue
+    let pendingDismissQueue: PendingNotificationDismissQueue
     private let pairingHintDefaults: UserDefaults
+    // Internal (not private): used by the dismiss-sync and dogfood-feedback
+    // extension files.
     let clientID: String
     /// Delivers the email path of Send Feedback (`/api/feedback`). `nil` when the
     /// web API base URL is unavailable; the email path then fails closed and the
@@ -451,6 +454,8 @@ public final class MobileShellComposite: MobileTerminalOutputSinking {
     /// `public` so the DEV feedback-submit affordance can ``DiagnosticLog/export()``
     /// it.
     public let diagnosticLog: DiagnosticLog?
+    // Internal (not private): used by the dismiss-sync and dogfood-feedback
+    // extension files.
     var remoteClient: MobileCoreRPCClient? {
         didSet {
             if remoteClient == nil {
@@ -815,240 +820,6 @@ public final class MobileShellComposite: MobileTerminalOutputSinking {
         } catch {
             mobileShellLog.error("click forward failed surface=\(surfaceID, privacy: .public) error=\(String(describing: error), privacy: .public)")
         }
-    }
-
-    // MARK: - Notification dismiss-sync
-
-    /// Tell the Mac that one or more mirrored notifications were dismissed on
-    /// this phone (a swipe/clear on the delivered banner). The Mac marks them
-    /// read and clears its own banner; its store then emits `notification.dismissed`
-    /// back, which is a harmless no-op for the already-removed phone banner.
-    ///
-    /// Fire-and-forget against the authoritative Mac store, but durable: the ids
-    /// are enqueued in ``PendingNotificationDismissQueue`` BEFORE the RPC is
-    /// attempted and removed only after it succeeds, so a dismiss that races a
-    /// dead/absent attach channel is retried by ``flushPendingNotificationDismisses()``
-    /// on the next successful (re)subscribe instead of being dropped.
-    ///
-    /// Carries only opaque notification UUIDs, never terminal content, so it is
-    /// safe regardless of the Mac's phone-forward hideContent setting.
-    /// - Parameter ids: The stable notification ids the user dismissed.
-    public func dismissNotification(ids: [String]) async {
-        let trimmed = ids
-            .map { $0.trimmingCharacters(in: .whitespacesAndNewlines) }
-            .filter { !$0.isEmpty }
-        guard !trimmed.isEmpty else { return }
-        pendingDismissQueue.enqueue(trimmed)
-        guard let client = remoteClient else { return }
-        do {
-            let request = try MobileCoreRPCClient.requestData(
-                method: "notification.dismiss",
-                params: [
-                    "notification_ids": trimmed,
-                    "client_id": clientID,
-                ]
-            )
-            _ = try await client.sendRequest(request)
-            pendingDismissQueue.remove(trimmed)
-        } catch {
-            // Left in the queue; the next successful (re)subscribe re-sends.
-            mobileShellLog.error("notification dismiss sync failed count=\(trimmed.count, privacy: .public) error=\(String(describing: error), privacy: .public)")
-        }
-    }
-
-    /// Re-send every dismissal still waiting in the durable outbox (a swipe that
-    /// happened while detached, backgrounded, or before any scene existed). The
-    /// Mac ignores ids it does not know, so replaying stale entries is harmless.
-    func flushPendingNotificationDismisses() async {
-        let pending = pendingDismissQueue.pendingIDs
-        guard !pending.isEmpty else { return }
-        await dismissNotification(ids: pending)
-    }
-
-    /// Clear delivered banners on this phone in response to a Mac-side dismiss
-    /// (`notification.dismissed` peer event). The ids are stable Mac-side
-    /// notification ids; the injected ``DeliveredNotificationClearing`` seam
-    /// maps them to the matching delivered banners via their
-    /// `cmux.notificationId` payload key.
-    /// - Parameter ids: The notification ids the Mac dismissed.
-    public func clearDeliveredNotifications(ids: [String]) async {
-        let trimmed = ids
-            .map { $0.trimmingCharacters(in: .whitespacesAndNewlines) }
-            .filter { !$0.isEmpty }
-        guard !trimmed.isEmpty else { return }
-        await deliveredNotificationClearer.removeDelivered(ids: trimmed)
-    }
-
-    /// SET the app-icon badge to the Mac's authoritative unread total. Always an
-    /// absolute write (never local +/-1 arithmetic) so any drift self-heals on
-    /// the next event, push, or reconcile sweep.
-    /// - Parameter count: The Mac's unread-notification count.
-    public func applyAuthoritativeUnreadBadge(_ count: Int) {
-        deliveredNotificationClearer.setBadgeCount(max(0, count))
-    }
-
-    /// Kick off one foreground/connect dismiss-sync pass against `client`:
-    /// first flush the durable phone→Mac dismiss outbox (so the Mac's store
-    /// reflects every swipe that happened while detached), then run the
-    /// reconcile sweep (lane 3) whose answer therefore already includes them.
-    /// Fire-and-forget; failures are non-fatal because the next (re)subscribe
-    /// runs another pass.
-    private func scheduleNotificationReconcile(client: MobileCoreRPCClient) {
-        Task { [weak self] in
-            await self?.flushPendingNotificationDismisses()
-            await self?.reconcileNotificationsWithMac(client: client)
-        }
-    }
-
-    /// The reconcile sweep: send the Mac the ids of every banner currently
-    /// delivered on this phone; it answers with the subset handled there (read,
-    /// or recently dismissed/removed) plus its authoritative unread count. Remove
-    /// the handled banners and SET the badge. This heals whatever the live event
-    /// lane and the budgeted silent-push lane missed while the app was closed.
-    /// Ids the Mac does not recognize are left alone (they may mirror a
-    /// different paired Mac). Exchanges only opaque ids and a count.
-    func reconcileNotificationsWithMac(client: MobileCoreRPCClient) async {
-        let deliveredIDs = await deliveredNotificationClearer.deliveredIdentifiers()
-        guard remoteClient === client, connectionState == .connected else { return }
-        do {
-            let request = try MobileCoreRPCClient.requestData(
-                method: "notification.reconcile",
-                params: [
-                    "delivered_ids": deliveredIDs,
-                    "client_id": clientID,
-                ]
-            )
-            let data = try await client.sendRequest(request)
-            guard remoteClient === client else { return }
-            let response = try MobileNotificationReconcileResponse.decode(data)
-            await applyNotificationReconcile(response)
-            MobileDebugLog.anchormux(
-                "notif.reconcile delivered=\(deliveredIDs.count) handled=\(response.handledIDs.count) unread=\(response.unreadCount.map(String.init) ?? "nil")"
-            )
-        } catch {
-            // Older Macs don't implement the verb (method_not_found), and
-            // transport hiccups are non-fatal: the next (re)subscribe retries.
-            MobileDebugLog.anchormux("notif.reconcile_failed error=\(error)")
-        }
-    }
-
-    /// Apply a reconcile result: clear the banners the Mac reports handled and
-    /// SET the badge to its authoritative count. Split from the transport so the
-    /// behavior is unit-testable through the injected clearing seam.
-    func applyNotificationReconcile(_ response: MobileNotificationReconcileResponse) async {
-        if !response.handledIDs.isEmpty {
-            await clearDeliveredNotifications(ids: response.handledIDs)
-        }
-        if let unreadCount = response.unreadCount {
-            applyAuthoritativeUnreadBadge(unreadCount)
-        }
-    }
-
-    /// Privileged direct-to-agent feedback round-trip: export the structured
-    /// diagnostic log, package it with the supplied debug-log text, visible
-    /// terminal text, and an optional freeform note, and submit it to the paired
-    /// Mac's `dogfood.feedback.submit` sink so the existing watcher under
-    /// `~/.cache/cmux-dogfood-feedback/` catches it.
-    ///
-    /// This is the privileged path of the Send Feedback feature: it is offered
-    /// only to `@manaflow.ai` users on an active mobile-host connection (see
-    /// ``MobileFeedbackRoute/resolve(email:hasActiveMacConnection:hostSupportsAgentSink:)``), and is NOT
-    /// `#if DEBUG`-gated, so it works on Release (beta/prod) builds for the team.
-    ///
-    /// The structured log is exported here (the store owns ``diagnosticLog``);
-    /// the string snapshots are gathered by the caller on the UI layer, where the
-    /// `GhosttySurfaceView`/`MobileDebugLog` accessors live. Fire-and-forget; a
-    /// transport failure is logged and surfaced via the returned `Bool`.
-    ///
-    /// - Parameters:
-    ///   - text: An optional freeform note from the user.
-    ///   - debugLogText: The string debug-log snapshot (from `MobileDebugLog`).
-    ///   - terminalText: The visible terminal text (from `GhosttySurfaceView`).
-    ///   - buildStamp: The build-identity stamp (build type + version + OS +
-    ///     device) written into the bundle. Defaults to the diagnostic log's
-    ///     stamp when not supplied.
-    /// - Returns: `true` when the Mac acknowledged the bundle.
-    @discardableResult
-    public func submitPrivilegedAgentFeedback(
-        text: String,
-        debugLogText: String,
-        terminalText: String,
-        buildStamp: String? = nil
-    ) async -> Bool {
-        guard let client = remoteClient else { return false }
-        let diagnosticBlob = await diagnosticLog?.export() ?? Data()
-        let buildStamp = buildStamp ?? diagnosticLog?.buildStamp ?? ""
-        let clientID = clientID
-        // Cap inputs and build the (potentially multi-MiB) combined blob +
-        // base64 + JSON request OFF the main actor: the store is `@MainActor`, so
-        // doing the concat/encode here would block the UI on a large bundle. A
-        // detached task returns the finished request bytes (`Data` is `Sendable`).
-        let request: Data?
-        do {
-            request = try await Task.detached(priority: .utility) { () -> Data in
-                try Self.buildDogfoodFeedbackRequest(
-                    text: text,
-                    debugLogText: debugLogText,
-                    terminalText: terminalText,
-                    buildStamp: buildStamp,
-                    clientID: clientID,
-                    diagnosticBlob: diagnosticBlob
-                )
-            }.value
-        } catch {
-            mobileShellLog.error("dogfood feedback encode failed error=\(String(describing: error), privacy: .public)")
-            return false
-        }
-        guard let request else { return false }
-        do {
-            _ = try await client.sendRequest(request)
-            return true
-        } catch {
-            mobileShellLog.error("dogfood feedback submit failed error=\(String(describing: error), privacy: .public)")
-            return false
-        }
-    }
-
-    /// Client-side caps mirroring the Mac sink, applied before any large
-    /// allocation so a huge debug log or note can't be encoded into a multi-MiB
-    /// request on the phone. `nonisolated` so the off-main request builder can
-    /// read them.
-    nonisolated private static let dogfoodFeedbackMaxTextChars = 16_384
-    nonisolated private static let dogfoodFeedbackMaxTerminalChars = 262_144
-    nonisolated private static let dogfoodFeedbackMaxDebugLogChars = 1_048_576
-
-    /// Combine the structured + string diagnostics into one self-contained blob,
-    /// base64-encode it, and build the RPC request — all off the main actor.
-    ///
-    /// The string debug log rides inside the same diagnostic file as the compact
-    /// structured rows (rows, a divider, then the human-readable log) so the Mac
-    /// bundle is self-contained. Inputs are size-capped first.
-    nonisolated private static func buildDogfoodFeedbackRequest(
-        text: String,
-        debugLogText: String,
-        terminalText: String,
-        buildStamp: String,
-        clientID: String,
-        diagnosticBlob: Data
-    ) throws -> Data {
-        let cappedText = String(text.prefix(dogfoodFeedbackMaxTextChars))
-        let cappedTerminal = String(terminalText.prefix(dogfoodFeedbackMaxTerminalChars))
-        let cappedDebugLog = String(debugLogText.prefix(dogfoodFeedbackMaxDebugLogChars))
-        var combined = diagnosticBlob
-        if !cappedDebugLog.isEmpty {
-            combined.append(Data("\n----- mobile debug log -----\n".utf8))
-            combined.append(Data(cappedDebugLog.utf8))
-        }
-        return try MobileCoreRPCClient.requestData(
-            method: "dogfood.feedback.submit",
-            params: [
-                "text": cappedText,
-                "terminal_text": cappedTerminal,
-                "build_stamp": buildStamp,
-                "diagnostic_blob_base64": combined.base64EncodedString(),
-                "client_id": clientID,
-            ]
-        )
     }
 
     // MARK: - Feedback routing
